# Reduce allocations in various functions