The problem areas in these functions were found with a combination of `@report_opt` from JET.jl and `@profview_allocs` from ProfileCanvas.jl.

Benchmarks:
```julia
@benchmark peakproms!(pks) setup=(pks = findmaxima([0,5,2,3,3,1,4,0]))
# master 2.8 μs; 18 allocs, 848 bytes
# PR min 337 ns; 4 allocs, 224 bytes

@benchmark peakproms!(pks; strict=false) setup=(pks = findmaxima([0,5,2,3,3,1,4,0]))
# master 2.98 μs; 26 allocs, 1.25 KiB
# PR min 625 ns; 9 allocs, 608 bytes

@benchmark peakwidths!(pks) setup=(pks = peakproms(findmaxima([0,5,2,3,3,1,4,0])))
# master 2.2 μs; 51 allocs, 1.25 KiB
# PR min 418 ns; 14 allocs, 624 bytes

@benchmark peakwidths!(pks; strict=false) setup=(pks = peakproms(findmaxima([0,5,2,3,3,1,4,0])))
# master 2.95 μs; 69 allocs, 1.53 KiB
# PR min 434 ns; 14 allocs, 624 bytes

@benchmark filterpeaks!(pks, :heights; min=1) setup=(pks = findmaxima([0,5,2,3,3,1,4,0]))
# master 264 ns; 8 allocs, 464 bytes
# PR min 64 ns; 1 allocs, 64 bytes
```